CloudBack/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==10.0.0
